CLEANUP: Remove incorrectly commited root-level frontend folder
