--- frontend/.env
@@ -1,2 +1,0 @@
-VITE_SUPABASE_URL=https://cynzgofcoywdhmjjniyr.supabase.co
-VITE_SUPABASE_ANON_KEY=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZSIsInJlZiI6ImN5bnpnb2Zjb3l3ZGhtampuaXlyIiwicm9sZSI6ImFub24iLCJpYXQiOjE3NDE3MzY2NTMsImV4cCI6MjA1NzMxMjY1M30.4x5S6AnppP24XwB8UKK1zlcjqNd0KiJlSQnVjjM5uME
