--- frontend/.vite/deps/_metadata.json
@@ -1,8 +1,0 @@
-{
-  "hash": "1f368386",
-  "configHash": "66b61e2c",
-  "lockfileHash": "e3b0c442",
-  "browserHash": "000bdb96",
-  "optimized": {},
-  "chunks": {}
-}
