--- frontend/src/App.tsx
@@ -1,43 +1,0 @@
-import { BrowserRouter, Routes, Route, Navigate } from 'react-router-dom';
-import DashboardLayout from './components/Layout/DashboardLayout';
-import ProtectedRoute from './components/ProtectedRoute';
-import Dashboard from './pages/Dashboard';
-import Courses from './pages/Courses';
-import Attendance from './pages/Attendance';
-import Settings from './pages/Settings';
-import Account from './pages/Account';
-import FaceRegistration from './pages/Account/FaceRegistration';
-
-// Auth Pages
-import Login from './pages/Auth/Login';
-import Signup from './pages/Auth/Signup';
-
-function App() {
-  return (
-    <BrowserRouter>
-      <Routes>
-        {/* Default Route - Redirect to Login */}
-        <Route path="/" element={<Navigate to="/login" replace />} />
-
-        {/* Public/Auth Routes */}
-        <Route path="/login" element={<Login />} />
-        <Route path="/signup" element={<Signup />} />
-
-        {/* Protected Routes */}
-        <Route element={<ProtectedRoute />}>
-          <Route element={<DashboardLayout />}>
-            <Route path="/dashboard" element={<Dashboard />} />
-            <Route path="/courses" element={<Courses />} />
-            <Route path="/attendance" element={<Attendance />} />
-            <Route path="/settings" element={<Settings />} />
-            <Route path="/account" element={<Account />} />
-            <Route path='/face-registration' element={<FaceRegistration />} />
-            <Route path="/" element={<Navigate to="/dashboard" replace />} />
-          </Route>
-        </Route>
-      </Routes>
-    </BrowserRouter>
-  );
-}
-
-export default App;
