--- frontend/src/components/Layout/DashboardLayout.tsx
@@ -1,22 +1,0 @@
-import React from 'react';
-import { Outlet } from 'react-router-dom';
-import Sidebar from './Sidebar';
-import TopBar from './TopBar';
-
-const DashboardLayout: React.FC = () => {
-  return (
-    <div className="min-h-screen bg-gray-50 flex">
-      <Sidebar />
-      <div className="lg:pl-72 flex flex-col w-full">
-        <TopBar />
-        <main className="flex-grow bg-[#f3f6fd]">
-          <div className="">
-            <Outlet />
-          </div>
-        </main>
-      </div>
-    </div>
-  );
-};
-
-export default DashboardLayout;
