--- frontend/src/main.tsx
@@ -1,11 +1,0 @@
-import { StrictMode } from 'react';
-import { createRoot } from 'react-dom/client';
-import App from './App.tsx';
-import './index.css';
-
-
-createRoot(document.getElementById('root')!).render(
-  <StrictMode>
-      <App />
-  </StrictMode>
-);
