--- frontend/src/types/auth.ts
@@ -1,11 +1,0 @@
-export interface User {
-  id: string;
-  name: string;
-  email: string;
-  profileImage?: string;
-}
-
-export interface AuthState {
-  isAuthenticated: boolean;
-  user: User | null;
-}
