--- frontend/tailwind.config.js
@@ -1,15 +1,0 @@
-/** @type {import('tailwindcss').Config} */
-export default {
-  content: [  
-    './index.html', 
-    './src/**/*.{js,ts,jsx,tsx}',
-    "./pages/**/*.{js,ts,jsx,tsx}",
-    "./components/**/*.{js,ts,jsx,tsx}",
-    "./app/**/*.{js,ts,jsx,tsx}",
-    "./node_modules/@heroui/react/**/*.{js,ts,jsx,tsx}",
-  ],
-  theme: {
-    extend: {},
-  },
-  plugins: [],
-};
